LayoutTests/imported/w3c/web-platform-tests/css/css-position/sticky/sticky-after-input-expected.txt


1
2
3
4
5
6
7
8
9
10
11
12
13
14
15
16
17

FAIL Sticky positioned element should reset the scroll position to unshifted position assert_equals: should reset the scroll to unshifted sticky position expected 0 but got 208
